Replaced MapQuest routing with CloudMade. Fixes #32. Since it probably doesn't apply anymore, let's also close #38. That lets us stick with OSM data only. This is good for 2 reasons: 1) OSM is the data TriMet is using and improving, and 2) it feels more open-sourcy.
We should continue to monitor the thread on the MQ API forum; I'm not particularly tied to CM over MQ or vice versa. Indeed, MQ's compressed line option is desirable. If we do go back to MQ, a3364109225402e129feea48b4b00f2c17203949 is the last commit with the MQ code in it (I removed it to save space). We'll also need to remember to add back the MQ attribution.